+++ sentry-core/src/main/java/io/sentry/core/MainEventProcessor.java
@@ -1,7 +1,6 @@
 package io.sentry.core;
 
-import io.sentry.core.hints.ApplyScopeData;
-import io.sentry.core.hints.Cached;
+import io.sentry.core.util.ApplyScopeUtils;
 import io.sentry.core.util.Objects;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
@@ -47,7 +46,7 @@ public final class MainEventProcessor implements EventProcessor {
       event.setExceptions(sentryExceptionFactory.getSentryExceptions(throwable));
     }
 
-    if (shouldApplyScopeData(hint)) {
+    if (ApplyScopeUtils.shouldApplyScopeData(hint)) {
       processNonCachedEvent(event);
     } else {
       options
@@ -78,9 +77,5 @@ public final class MainEventProcessor implements EventProcessor {
     if (event.getThreads() == null && options.isAttachThreads()) {
       event.setThreads(sentryThreadFactory.getCurrentThreads());
     }
-  }
-
-  private boolean shouldApplyScopeData(final @Nullable Object hint) {
-    return (!(hint instanceof Cached) || (hint instanceof ApplyScopeData));
   }
 }
