+++ sentry-core/src/main/java/io/sentry/core/SentryClient.java
@@ -4,13 +4,12 @@ import io.sentry.core.cache.DiskCache;
 import io.sentry.core.cache.IEnvelopeCache;
 import io.sentry.core.cache.IEventCache;
 import io.sentry.core.cache.SessionCache;
-import io.sentry.core.hints.ApplyScopeData;
-import io.sentry.core.hints.Cached;
 import io.sentry.core.hints.SessionUpdateHint;
 import io.sentry.core.protocol.SentryId;
 import io.sentry.core.transport.Connection;
 import io.sentry.core.transport.ITransport;
 import io.sentry.core.transport.ITransportGate;
+import io.sentry.core.util.ApplyScopeUtils;
 import io.sentry.core.util.Objects;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -66,10 +65,6 @@ public final class SentryClient implements ISentryClient {
     random = options.getSampleRate() == null ? null : new Random();
   }
 
-  private boolean shouldApplyScopeData(final @Nullable Object hint) {
-    return (!(hint instanceof Cached) || (hint instanceof ApplyScopeData));
-  }
-
   @Override
   public @NotNull SentryId captureEvent(
       @NotNull SentryEvent event, final @Nullable Scope scope, final @Nullable Object hint) {
@@ -77,7 +72,7 @@ public final class SentryClient implements ISentryClient {
 
     options.getLogger().log(SentryLevel.DEBUG, "Capturing event: %s", event.getEventId());
 
-    if (shouldApplyScopeData(hint)) {
+    if (ApplyScopeUtils.shouldApplyScopeData(hint)) {
       // Event has already passed through here before it was cached
       // Going through again could be reading data that is no longer relevant
       // i.e proguard id, app version, threads
@@ -151,7 +146,7 @@ public final class SentryClient implements ISentryClient {
   @TestOnly
   void updateSessionData(
       final @NotNull SentryEvent event, final @Nullable Object hint, final @Nullable Scope scope) {
-    if (shouldApplyScopeData(hint)) {
+    if (ApplyScopeUtils.shouldApplyScopeData(hint)) {
       // safe guard
       if (options.isEnableSessionTracking()) {
         if (scope != null) {
