+++ sentry-core/src/main/java/io/sentry/core/hints/ApplyScopeData.java
@@ -1,4 +1,7 @@
 package io.sentry.core.hints;
 
-/** Marker interface for applying scope's data */
+/**
+ * Marker interface for applying scope's data. This means applying data relevant to the current
+ * execution like applying threads or current app version.
+ */
 public interface ApplyScopeData {}
