+++ sentry-core/src/main/java/io/sentry/core/hints/Cached.java
@@ -1,6 +1,4 @@
 package io.sentry.core.hints;
 
-// Marker interface for a capture involving data cached from disk
-// This means applying data relevant to the current execution should be avoided
-// like applying threads or current app version.
+/** Marker interface for a capture involving data cached from disk */
 public interface Cached {}
