--- sentry-core/src/main/java/io/sentry/core/hints/Resettable.java
@@ -1,7 +1,0 @@
-package io.sentry.core.hints;
-
-/** marker interface that resets/restarts a Flushable operation. */
-public interface Resettable {
-
-  void reset();
-}
