--- sentry-core/src/main/java/io/sentry/core/hints/SendCachedEventHint.java
@@ -1,57 +1,0 @@
-package io.sentry.core.hints;
-
-import static io.sentry.core.SentryLevel.ERROR;
-
-import io.sentry.core.ILogger;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
-
-@ApiStatus.Internal
-public final class SendCachedEventHint implements Cached, Retryable, SubmissionResult, Flushable {
-  boolean retry = false;
-  boolean succeeded = false;
-
-  private final CountDownLatch latch;
-  private final long flushTimeoutMillis;
-  private final @NotNull ILogger logger;
-
-  public SendCachedEventHint(final long flushTimeoutMillis, final @NotNull ILogger logger) {
-    this.flushTimeoutMillis = flushTimeoutMillis;
-    this.latch = new CountDownLatch(1);
-    this.logger = logger;
-  }
-
-  @Override
-  public boolean isRetry() {
-    return retry;
-  }
-
-  @Override
-  public void setRetry(boolean retry) {
-    this.retry = retry;
-  }
-
-  @Override
-  public boolean waitFlush() {
-    try {
-      return latch.await(flushTimeoutMillis, TimeUnit.MILLISECONDS);
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-      logger.log(ERROR, "Exception while awaiting on lock.", e);
-    }
-    return false;
-  }
-
-  @Override
-  public void setResult(boolean succeeded) {
-    this.succeeded = succeeded;
-    latch.countDown();
-  }
-
-  @Override
-  public boolean isSuccess() {
-    return succeeded;
-  }
-}
